Reword router explanation
The canonical example now routes to controllers, not endpoints, so "to other
wsgi apps" no longer fits here.

I would like to expand my docs and explain exactly how the router works, and
that it technically is routing to other wsgi apps, and that the `add` method
is just a convenience method that wraps your controller in an Endpoint (which
is a wsgi app).